Inlining in container
